requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
openai>=1.10.0
pydantic>=2.7.4
